#21811 - Rest-API updating product stock_item deletes downloadable_product_links

+++ app/code/Magento/Downloadable/Model/Link/UpdateHandler.php
@@ -44,16 +44,18 @@ class UpdateHandler implements ExtensionInterface
         $links = $entity->getExtensionAttributes()->getDownloadableProductLinks() ?: [];
         $updatedLinks = [];
         $oldLinks = $this->linkRepository->getList($entity->getSku());
-        foreach ($links as $link) {
-            if ($link->getId()) {
-                $updatedLinks[$link->getId()] = true;
+        if (!empty($links)) {
+            foreach ($links as $link) {
+                if ($link->getId()) {
+                    $updatedLinks[$link->getId()] = true;
+                }
+                $this->linkRepository->save($entity->getSku(), $link, !(bool)$entity->getStoreId());
             }
-            $this->linkRepository->save($entity->getSku(), $link, !(bool)$entity->getStoreId());
-        }
-        /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
-        foreach ($oldLinks as $link) {
-            if (!isset($updatedLinks[$link->getId()])) {
-                $this->linkRepository->delete($link->getId());
+            /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
+            foreach ($oldLinks as $link) {
+                if (!isset($updatedLinks[$link->getId()])) {
+                    $this->linkRepository->delete($link->getId());
+                }
             }
         }
